Move test bad_filter.rs to api_find.rs.
